Upgrade to Groovy 5

+++ matrix-spreadsheet/src/main/groovy/se/alipsa/matrix/spreadsheet/fastexcel/FExcelReader.groovy
@@ -169,7 +169,7 @@ class FExcelReader implements SpreadsheetReader {
   }
 
   static int findLastRow(Sheet sheet) {
-    return sheet.openStream().count()
+    return sheet.openStream().count() as int
   }
 
   @Override
